Move around files to tidy repo
