+++ scp/run_simulation_scp_nn.py
